+++ integreat_cms/cms/models/pois/poi_translation.py
@@ -88,4 +88,4 @@ class POITranslation(AbstractContentTranslation):
         #: The default permissions for this model
         default_permissions = ()
         #: The fields which are used to sort the returned objects of a QuerySet
-        ordering = ["poi__pk", "-version"]
+        ordering = ["poi__pk", "language__pk", "-version"]